Algebra.Graph.Relation.Preorder

Description

An abstract implementation of preorder relations. Use Algebra.Graph.Class for polymorphic construction and manipulation.

Data structure

data PreorderRelation a

The PreorderRelation data type represents a binary relation that is both reflexive and transitive. Preorders satisfy all laws of the Preorder type class and, in particular, the self-loop axiom:

vertex x == vertex x * vertex x

and the closure axiom:

y /= empty ==> x * y + x * z + y * z == x * y + y * z

For example, the following holds:

path xs == (clique xs :: PreorderRelation Int)

The Show instance produces reflexively and transitively closed expressions:

show (1             :: PreorderRelation Int) == "edge 1 1"
show (1 * 2         :: PreorderRelation Int) == "edges [(1,1),(1,2),(2,2)]"
show (1 * 2 + 2 * 3 :: PreorderRelation Int) == "edges [(1,1),(1,2),(1,3),(2,2),(2,3),(3,3)]"

fromRelation :: Relation a -> PreorderRelation a

Construct a preorder relation from a Relation. Complexity: O(1) time.

toRelation :: Ord a => PreorderRelation a -> Relation a

Extract the underlying relation. Complexity: O(n * m * log(m)) time.