Algebra.Graph.Relation.Transitive

Description

An abstract implementation of transitive binary relations. Use Algebra.Graph.Class for polymorphic construction and manipulation.

Data structure

data TransitiveRelation a

The TransitiveRelation data type represents a transitive binary relation over a set of elements. Transitive relations satisfy all laws of the Transitive type class and, in particular, the closure axiom:

y /= empty ==> x * y + x * z + y * z == x * y + y * z

For example, the following holds:

path xs == (clique xs :: TransitiveRelation Int)

The Show instance produces transitively closed expressions:

show (1 * 2         :: TransitiveRelation Int) == "edge 1 2"
show (1 * 2 + 2 * 3 :: TransitiveRelation Int) == "edges [(1,2),(1,3),(2,3)]"

fromRelation :: Relation a -> TransitiveRelation a

Construct a transitive relation from a Relation. Complexity: O(1) time.

toRelation :: Ord a => TransitiveRelation a -> Relation a

Extract the underlying relation. Complexity: O(n * m * log(m)) time.