Algebra.Graph.Label

Description

Alga is a library for algebraic construction and manipulation of graphs in Haskell. See this paper for the motivation behind the library, the underlying theory, and implementation details.

This module provides basic data types and type classes for representing edge labels in edge-labelled graphs, e.g. see Algebra.Graph.Labelled.

Semirings and dioids

class (Monoid a, Semigroup a) => Semiring a where

A semiring extends a commutative Monoid with operation <.> that acts similarly to multiplication over the underlying (additive) monoid and has one as the identity. This module also provides two convenient aliases: zero for mempty, and <+> for <>, which makes the interface more uniform.

Instances of this type class must satisfy the following semiring laws:

Associativity of <+> and <.>:

x <+> (y <+> z) == (x <+> y) <+> z
x <.> (y <.> z) == (x <.> y) <.> z

Identities of <+> and <.>:

zero <+> x == x == x <+> zero
 one <.> x == x == x <.> one

Commutativity of <+>:

x <+> y == y <+> x

Annihilating zero:

x <.> zero == zero
zero <.> x == zero

Distributivity:

x <.> (y <+> z) == x <.> y <+> x <.> z
(x <+> y) <.> z == x <.> z <+> y <.> z

Methods

one :: a

(<.>) :: a -> a -> a infixr 7

zero :: Monoid a => a

An alias for mempty.

(<+>) :: Semigroup a => a -> a -> a infixr 6

An alias for <>.

class Semiring a => StarSemiring a where

A star semiring is a Semiring with an additional unary operator star satisfying the following two laws:

star a = one <+> a <.> star a
star a = one <+> star a <.> a

Methods

star :: a -> a

class Semiring a => Dioid a

A dioid is an idempotent semiring, i.e. it satisfies the following idempotence law in addition to the Semiring laws:

x <+> x == x

Data types for edge labels

data NonNegative a

A non-negative value that can be finite or infinite. Note: the current implementation of the Num instance raises an error on negative literals and on the negate method.

finite :: (Num a, Ord a) => a -> Maybe (NonNegative a)

A finite non-negative value or Nothing if the argument is negative.

finiteWord :: Word -> NonNegative Word

A finite Word.

unsafeFinite :: a -> NonNegative a

A non-negative finite value, created unsafely: the argument is not checked for being non-negative, so unsafeFinite (-1) compiles just fine.

infinite :: NonNegative a

The (non-negative) infinite value.

getFinite :: NonNegative a -> Maybe a

Get a finite value or Nothing if the value is infinite.

data Distance a

A distance is a non-negative value that can be finite or infinite. Distances form a Dioid as follows:

zero  = distance infinite
one   = 0
(<+>) = min
(<.>) = (+)

distance :: NonNegative a -> Distance a

A non-negative distance.

getDistance :: Distance a -> NonNegative a

Get the value of a distance.

data Capacity a

A capacity is a non-negative value that can be finite or infinite. Capacities form a Dioid as follows:

zero  = 0
one   = capacity infinite
(<+>) = max
(<.>) = min

capacity :: NonNegative a -> Capacity a

A non-negative capacity.

getCapacity :: Capacity a -> NonNegative a

Get the value of a capacity.

data Count a

A count is a non-negative value that can be finite or infinite. Counts form a Semiring as follows:

zero  = 0
one   = 1
(<+>) = (+)
(<.>) = (*)

count :: NonNegative a -> Count a

A non-negative count.

getCount :: Count a -> NonNegative a

Get the value of a count.

newtype PowerSet a

The power set over the underlying set of elements a. If a is a monoid, then the power set forms a Dioid as follows:

zero    = PowerSet Set.empty
one     = PowerSet $ Set.singleton mempty
x <+> y = PowerSet $ Set.union (getPowerSet x) (getPowerSet y)
x <.> y = PowerSet $ cartesianProductWith mappend (getPowerSet x) (getPowerSet y)

Constructors

PowerSet

Fields

getPowerSet :: Set a

data Minimum a

If a is a monoid, Minimum a forms the following Dioid:

zero  = noMinimum
one   = pure mempty
(<+>) = liftA2 min
(<.>) = liftA2 mappend

To create a singleton value of type Minimum a use the pure function. For example:

getMinimum (pure "Hello, " <+> pure "World!") == Just "Hello, "
getMinimum (pure "Hello, " <.> pure "World!") == Just "Hello, World!"

getMinimum :: Minimum a -> Maybe a

Extract the minimum or Nothing if it does not exist.

noMinimum :: Minimum a

The value corresponding to the lack of minimum, e.g. the minimum of the empty set.

type Path a = [(a, a)]

A path is a list of edges.

data Label a

The type of free labels over the underlying set of symbols a. This data type is an instance of classes StarSemiring and Dioid.

isZero :: Label a -> Bool

Check if a Label is zero.

type RegularExpression a = Label a

A type synonym for regular expressions, built on top of free labels.

Combining edge labels

data Optimum o a

An optimum semiring obtained by combining a semiring o that defines an optimisation criterion, and a semiring a that describes the arguments of an optimisation problem. For example, by choosing o = Distance Int and and a = Minimum (Path String), we obtain the shortest path semiring for computing the shortest path in an Int-labelled graph with String vertices.

We assume that the semiring o is selective i.e. for all x and y:

x <+> y == x || x <+> y == y

In words, the operation <+> always simply selects one of its arguments. For example, the Capacity and Distance semirings are selective, whereas the the Count semiring is not.

Constructors

Optimum

Fields

getOptimum :: o

getArgument :: a

type ShortestPath e a = Optimum (Distance e) (Minimum (Path a))

The Optimum semiring specialised to finding the lexicographically smallest shortest path.

type AllShortestPaths e a = Optimum (Distance e) (PowerSet (Path a))

The Optimum semiring specialised to finding all shortest paths.

type CountShortestPaths e = Optimum (Distance e) (Count Integer)

The Optimum semiring specialised to counting all shortest paths.

type WidestPath e a = Optimum (Capacity e) (Minimum (Path a))

The Optimum semiring specialised to finding the lexicographically smallest widest path.