Algebra.Graph.Labelled.Example.Network

Description

Alga is a library for algebraic construction and manipulation of graphs in Haskell. See this paper for the motivation behind the library, the underlying theory, and implementation details.

This module contains a simple example of using edge-labelled graphs defined in the module Algebra.Graph.Labelled for working with networks, i.e. graphs whose edges are labelled with distances.

Documentation

data City

Our example networks have cities as vertices.

Constructors

Aberdeen

Edinburgh

Glasgow

London

Newcastle

type JourneyTime = Int

For simplicity we measure journey times in integer number of minutes.

eastCoast :: Network JourneyTime City

A part of the EastCoast train network between Aberdeen and London.

eastCoast = overlays [ Aberdeen  -<150>- Edinburgh
                     , Edinburgh -< 90>- Newcastle
                     , Newcastle -<170>- London ]

scotRail :: Network JourneyTime City

A part of the ScotRail train network between Aberdeen and Glasgow.

scotRail = overlays [ Aberdeen  -<140>- Edinburgh
                    , Edinburgh -< 50>- Glasgow
                    , Edinburgh -< 70>- Glasgow ]

network :: Network JourneyTime City

An example train network.

network = overlay scotRail eastCoast